requirements.txt
Django==2.1.4
django-cors-middleware==1.3.1
django-extensions==2.1.4
djangorestframework==3.9.0
PyJWT==1.7.1
pytz==2018.7
six==1.12.0
